Correctly merge the file field configuration
Currently it picks the settings in the wrong way, and even after fixing
it, it merges the config (adding multiple and destination as arrays, so
the code following will break as it assumes those are strings). The new
code replaces the default values with the array values set in the field.