Replace `display:none` by `visibility:hidden` & a hiding content technique:
The inline CSS `display:none` is useless because "turns off the
display of an element so that it has no effect on layout (the document
is rendered as though the element did not exist)." - Source
<https://developer.mozilla.org/en-US/docs/Web/CSS/display#display-box>

So the spambot can't fill it too.

The honeypot input must not be visible by human nor screen reader but
it should be focusable by a bot.