Fix checkboxes checked state
This is a long standing issue (7 years) for checkbox default values when `use: keys` is **not** used. The default values are still specified in a named array `defaultvalue => keyname` but the code assumes a flat array. When setting a default value of `false` the code will fail. You can test this using:

```yaml
my_field:
    type: checkboxes
    label: A couple of checkboxes
    default:
        option1: true
        option2: false
    options:
        option1: Option 1
        option2: Option 2
```